data-engineer-challenge/iot_ingest/requirements.txt
paho-mqtt==1.6.1
colorlog